+++ app/charts/map/chart-map.tsx
@@ -1,3 +1,4 @@
+import { Box } from "@mui/material";
 import { memo, useCallback, useMemo } from "react";
 
 import { ChartDataWrapper } from "@/charts/chart-data-wrapper";
@@ -6,13 +7,17 @@ import { MapComponent } from "@/charts/map/map";
 import { MapLegend } from "@/charts/map/map-legend";
 import { MapChart } from "@/charts/map/map-state";
 import { MapTooltip } from "@/charts/map/map-tooltip";
+import { MapWMTSLegend } from "@/charts/map/map-wmts-legend";
 import {
   ChartContainer,
   ChartControlsContainer,
 } from "@/charts/shared/containers";
 import { NoGeometriesHint } from "@/components/hint";
 import { Cube, MapConfig } from "@/config-types";
-import { useChartConfigFilters } from "@/config-utils";
+import {
+  useChartConfigFilters,
+  useDefinitiveTemporalFilterValue,
+} from "@/config-utils";
 import { TimeSlider } from "@/configurator/interactive-filters/time-slider";
 import {
   dimensionValuesToGeoCoordinates,
@@ -181,11 +186,15 @@ const ChartMap = memo((props: ChartMapProps) => {
   const { chartConfig, dimensions, observations } = props;
   const { fields } = chartConfig;
   const filters = useChartConfigFilters(chartConfig);
+  const temporalFilterValue = useDefinitiveTemporalFilterValue({ dimensions });
 
   return (
     <MapChart {...props}>
       <ChartContainer>
-        <MapComponent />
+        <MapComponent
+          customWMTSLayers={chartConfig.baseLayer.customWMTSLayers}
+          value={temporalFilterValue ? +temporalFilterValue : undefined}
+        />
         <MapTooltip />
       </ChartContainer>
       <ChartControlsContainer sx={{ mt: 6 }}>
@@ -196,7 +205,20 @@ const ChartMap = memo((props: ChartMapProps) => {
             {...fields.animation}
           />
         )}
-        <MapLegend chartConfig={chartConfig} observations={observations} />
+        <Box
+          sx={{
+            display: "flex",
+            flexDirection: "row",
+            gap: 4,
+            flexWrap: "wrap",
+          }}
+        >
+          <MapLegend chartConfig={chartConfig} observations={observations} />
+          <MapWMTSLegend
+            chartConfig={chartConfig}
+            value={temporalFilterValue ? +temporalFilterValue : undefined}
+          />
+        </Box>
       </ChartControlsContainer>
     </MapChart>
   );