+++ app/charts/shared/chart-state.ts
@@ -204,9 +204,9 @@ export const useBandYVariables = (
     getYLabel,
     getYAbbreviationOrLabel,
     yTimeUnit,
-    getYAsDate: isTemporalDimension(yDimension)
-      ? getYAsDate
-      : getYTemporalEntity,
+    getYAsDate: isTemporalEntityDimension(yDimension)
+      ? getYTemporalEntity
+      : getYAsDate,
   };
 };
 
@@ -249,9 +249,9 @@ export const useBandXVariables = (
     getXLabel,
     getXAbbreviationOrLabel,
     xTimeUnit,
-    getXAsDate: isTemporalDimension(xDimension)
-      ? getXAsDate
-      : getXTemporalEntity,
+    getXAsDate: isTemporalEntityDimension(xDimension)
+      ? getXTemporalEntity
+      : getXAsDate,
   };
 };