+++ app/configurator/components/field.tsx
@@ -37,7 +37,6 @@ import {
   Slider,
   Switch,
 } from "@/components/form";
-import { OpenMetadataPanelWrapper } from "@/components/metadata-panel";
 import SelectTree from "@/components/select-tree";
 import useDisclosure from "@/components/use-disclosure";
 import {
@@ -420,15 +419,7 @@ export const DataFilterTemporal = ({
     <>
       <DatePickerField
         name={`date-picker-${dimension.id}`}
-        label={
-          <FieldLabel
-            label={
-              <OpenMetadataPanelWrapper component={dimension}>
-                {label}
-              </OpenMetadataPanelWrapper>
-            }
-          />
-        }
+        label={<FieldLabel label={label} />}
         value={
           usesMostRecentDate ? maxDate : (parseDate(fieldProps.value) as Date)
         }